prod/requirements.txt
streamlit
torch
torchvision
scikit-learn
pillow

